@bott/storage module (wip)
