+++ app/commands/music.ts
@@ -1,4 +1,4 @@
-import { BottEventType } from "@bott/data";
+import { BottEventType } from "@bott/model";
 import { CommandOptionType, createCommand, createTask } from "@bott/discord";
 import { generateMusicFile } from "@bott/gemini";
 import { RATE_LIMIT_MUSIC, RATE_LIMIT_WINDOW_MS } from "../constants.ts";
