+++ app/commands/video.ts
@@ -1,4 +1,4 @@
-import { BottEventType } from "@bott/data";
+import { BottEventType } from "@bott/model";
 import { CommandOptionType, createCommand, createTask } from "@bott/discord";
 import { generateVideoFile } from "@bott/gemini";
