+++ app/deno.jsonc
@@ -4,7 +4,7 @@
     "start": "deno run --allow-all --env-file=.env.development ./main.ts"
   },
   "imports": {
-    "@bott/data": "../data/main.ts",
+    "@bott/model": "../data/main.ts",
     "@bott/gemini": "../helpers/gemini/main.ts",
     "@bott/discord": "../helpers/discord/main.ts"
   }
