--- data/deno.json
@@ -1,4 +1,0 @@
-{
-  "name": "@bott/data",
-  "exports": "./main.ts"
-}
