--- data/main.ts
@@ -1,17 +1,0 @@
-export { initDatabase } from "./database/commit.ts";
-export { setSchema } from "./model/schema.ts";
-export { type BottSpace } from "./model/spaces.ts";
-export { type BottChannel } from "./model/channels.ts";
-export { type BottUser } from "./model/users.ts";
-export {
-  type BottFile,
-  BottFileType,
-  getFileFromUrl,
-} from "./model/files/sql.ts";
-export {
-  addEvents,
-  type BottEvent,
-  BottEventType,
-  getEventIdsForChannel,
-  getEvents,
-} from "./model/events.ts";
