--- data/model/schema.ts
@@ -1,15 +1,0 @@
-import { commit } from "../database/commit.ts";
-import { channelsTableSql } from "./channels.ts";
-import { eventsTableSql } from "./events.ts";
-import { filesTableSql } from "./files/sql.ts";
-import { spacesTableSql } from "./spaces.ts";
-import { usersTableSql } from "./users.ts";
-
-export const setSchema = () =>
-  commit(
-    spacesTableSql,
-    channelsTableSql,
-    usersTableSql,
-    eventsTableSql,
-    filesTableSql,
-  );
