--- helpers/discord/bot/start.test.ts
@@ -1,1 +1,0 @@
-// TODO(#2): tests
