--- helpers/discord/embed/error.test.ts
@@ -1,1 +1,0 @@
-// TODO(#2): tests
