--- helpers/discord/embed/info.test.ts
@@ -1,1 +1,0 @@
-// TOOD: test
