+++ helpers/gemini/files/photo.ts
@@ -1,7 +1,7 @@
 import { PersonGeneration, SafetyFilterLevel } from "npm:@google/genai";
 import { decodeBase64 } from "jsr:@std/encoding";
 
-import { BottFileType } from "@bott/data";
+import { BottAssetType } from "@bott/model";
 
 import _gemini from "../client.ts";
 import type { FileGenerator } from "./types.ts";
@@ -53,6 +53,6 @@ export const generatePhotoFile: FileGenerator = async (prompt: string, {
     name: fileName,
     description: prompt,
     url: getGeneratedFileUrl(fileName),
-    type: BottFileType.PNG,
+    type: BottAssetType.PNG,
   };
 };
