+++ helpers/storage/assets/add.ts
@@ -1,7 +1,11 @@
-import { writeData } from "../../../filesystem/client.ts";
-import { type BottFile, BottFileType } from "../types.ts";
-import { BottFileSourceType } from "./types.ts";
-import { prepareHtml } from "./html.ts";
+import { join } from "jsr:@std/path";
+
+import { FS_ASSET_ROOT } from "../start.ts";
+
+import { type BottAsset, BottAssetType } from "@bott/model";
+
+import { SupportedRawFileType } from "./types.ts";
+import { prepareHtml } from "./prepare/html.ts";
 
 export const _getResponseContentType = (response: Response): string => {
   const contentTypeHeader = response.headers.get("content-type");
@@ -9,7 +13,13 @@ export const _getResponseContentType = (response: Response): string => {
   return contentTypeHeader.split(";")[0].trim();
 };
 
-export const prepareFileSource = async (source: URL): Promise<BottFile> => {
+export const addAsset = async (source: URL): Promise<BottAsset> => {
+  if (!FS_ASSET_ROOT) {
+    throw new Error(
+      "Storage has not been started: FS_ASSET_ROOT is not defined",
+    );
+  }
+
   // 1. resolve source URL
   const response = await fetch(source);
   const sourceData = new Uint8Array(await response.arrayBuffer());
@@ -18,7 +28,7 @@ export const prepareFileSource = async (source: URL): Promise<BottFile> => {
   // 2. prepare file of type
   let resultData, resultType;
   switch (sourceType) {
-    case BottFileSourceType.HTML:
+    case SupportedRawFileType.HTML:
       [resultData, resultType] = await prepareHtml(sourceData);
       break;
     default:
@@ -29,7 +39,7 @@ export const prepareFileSource = async (source: URL): Promise<BottFile> => {
   let path = resultType as string;
   let name = source.pathname.split("/").pop() || "index";
 
-  for (const [key, value] of Object.entries(BottFileType)) {
+  for (const [key, value] of Object.entries(BottAssetType)) {
     if (resultType === value) {
       name += `.${key.toLowerCase()}`;
       break;
@@ -38,9 +48,9 @@ export const prepareFileSource = async (source: URL): Promise<BottFile> => {
 
   path += `/${name}`;
 
-  writeData(resultData, path);
+  Deno.writeFileSync(join(FS_ASSET_ROOT, path), resultData);
 
-  // 4. return BottFile
+  // 4. return BottAsset
   return {
     id: crypto.randomUUID(),
     path,
