+++ model/deno.json
@@ -1,0 +1,4 @@
+{
+  "name": "@bott/model",
+  "exports": "./types.ts"
+}
